Committing a few fixes
git-svn-id: https://svn.code.sf.net/p/mbrt/code/trunk@112 a272ff57-b438-0410-9c92-d377374fc898

+++ plugins/checkermaterial.cpp
@@ -5,9 +5,9 @@ CheckeredMaterial::StaticInit CheckeredMaterial::m_init;
 
 int CheckeredMaterial::choose_material(const Point3D& intersection_point) const
 {
-    int x = (int) ( floor(intersection_point.x)) * m_scale ;
-    int y = (int) ( floor(intersection_point.y)) * m_scale ;
-    int z = (int) ( floor(intersection_point.z)) * m_scale ;
+    int x = (int)((floorl(intersection_point.x)) * m_scale);
+    int y = (int)((floorl(intersection_point.y)) * m_scale);
+    int z = (int)((floorl(intersection_point.z)) * m_scale);
 
     // -3 -2 -1  0  1  2  3
